pyproject.toml
[project]
name = "better-dynamic-contrast-ratio"
version = "0.1.0"
description = "Software implementation of Dynamic Contrast Ratio (DCR) for SDR monitors via DDC/CI"
authors = [{ name = "Danyal Zia Khan", email = "danyalziakhan@gmail.com" }]
requires-python = ">=3.11,<3.14"
readme = "README.md"
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Video :: Display",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: System :: Hardware",
]
dependencies = [
    "mss>=10.1.0",
    "numba>=0.64.0",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "zbl>=0.7.1",
]

[project.urls]
Homepage = "https://github.com/danyalziakhan/better-dynamic-contrast-ratio"
"Bug Tracker" = "https://github.com/danyalziakhan/better-dynamic-contrast-ratio/issues"

[dependency-groups]
dev = ["pre-commit>=4.5.1", "ruff>=0.15.7"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
managed = true

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["main.py", "config.py"]
exclude = ["**/node_modules", "**/__pycache__", "**/.venv", "**/.*"]
venvPath = "."
venv = ".venv"
pythonVersion = "3.13"
pythonPlatform = "Windows"

reportImportCycles = true
reportMissingImports = true
reportMissingTypeStubs = false
reportPrivateUsage = true
reportUnknownArgumentType = true
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = "warning"
reportUnknownVariableType = "warning"
reportUntypedFunctionDecorator = false
reportUnusedImport = "warning"
